requirements.txt
Flask==0.10.1
docker-py==0.3.1
Flask-WTF==0.9.5
